requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pygrib==2.1.4
numpy==1.26.4
matplotlib==3.8.3
cartopy==0.22.0
pandas==2.2.1
python-multipart==0.0.9
pillow==10.2.0
pydantic==2.6.1
requests==2.31.0
pytest==8.0.2
httpx==0.26.0 